+++ CHANGELOG.md
@@ -33,9 +33,9 @@
   [woxtu](https://github.com/woxtu)  
   [SimplyDanny](https://github.com/SimplyDanny)  
 
-* Add new `prefer_key_path` rule that triggers when a trailing closure on a function 
-  call is only hosting a (chained) member access expression since the closure can be
-  replaced with a key path argument.  
+* Add new `prefer_key_path` rule that triggers when a trailing closure on a standard
+  function call is only hosting a (chained) member access expression since the closure
+  can be replaced with a key path argument. Likewise, it triggers on closure arguments.  
   [SimplyDanny](https://github.com/SimplyDanny)
 
 * Adds `baseline` and `write_baseline` configuration file settings, equivalent